rag/requirements.txt
requests>=2.31.0
numpy>=1.26.0
PyYAML>=6.0.1
fastapi>=0.110.0
uvicorn>=0.27.0
